notes toward a mocha test, but it is too difficult because of incompatibilities between front and back end
