+++ lib/big-upload-client.js
@@ -1,5 +1,7 @@
-// Upload files in 4MB chunks
-const chunkSize = 1024; // 1024 * 1024 * 4;
+// Upload files in 4MB chunks, well under the typical
+// proxy server limit on POST size
+
+const defaultChunkSize = 1024 * 1024 * 4;
 
 // Used to upload very large files, large enough that they
 // might exceed the max body size or max uploaded file size
@@ -15,7 +17,7 @@ const chunkSize = 1024; // 1024 * 1024 * 4;
 //
 // An existing query string in `url` is NOT supported. Use `qs`.
 //
-// The receiving route must use the `self.apos.http.bigUpload`
+// The receiving route must use the `self.apos.http.bigUploadMiddleware`
 // function to obtain the middleware and it must be a `post` route.
 //
 // The query parameter `aposBigUpload` is reserved for internal use.
@@ -24,17 +26,26 @@ const chunkSize = 1024; // 1024 * 1024 * 4;
 // with a proportion (between 0.0 and 1.0) as the upload progresses.
 // Be aware the receiving end will typically do quite a bit of processing
 // after that.
+//
+// If `chunkSize` is specified (in bytes) it is used instead of the usual
+// 4MB. The chunk size should be big enough for efficiency but small
+// enough to avoid proxy server POST limits.
+//
+// For testing and server-side use an `http` object and a `FormData` constructor
+// may be specified otherwise the standard Apostrophe `apos.http` object
+// and the browser's `FormData` are used.
 
-export default async (url, options) => {
+module.exports = async (url, options) => {
 
-  const http = apos.http;
-
+  const chunkSize = options.chunkSize || defaultChunkSize;
+  const http = options.http || apos.http;
+  const File = options.File || window.File;
+  const FormData = options.FormData || window.FormData;
   const progress = options.progress || (n => {});
   const files = options.files || [];
   const info = {};
   let totalBytes = 0;
   let sentBytes = 0;
-  console.log('files:', files);
   for (const [ param, file ] of Object.entries(files)) {
     totalBytes += file.size;
     info[param] = {
@@ -43,7 +54,6 @@ export default async (url, options) => {
       chunks: Math.ceil(file.size / chunkSize)
     };
   }
-  console.log('info:', info);
   const { id } = await http.post(url, {
     qs: {
       aposBigUpload: {
@@ -54,7 +64,6 @@ export default async (url, options) => {
       files: info
     }
   });
-  console.log('after start');
   let n = 0;
   for (const [ param, file ] of Object.entries(files)) {
     const { name, size } = file;
@@ -63,8 +72,7 @@ export default async (url, options) => {
       const formData = new FormData();
       const thisChunkSize = Math.min(chunkSize, size - offset);
       formData.append('chunk', file.slice(offset, offset + thisChunkSize));
-      console.log('before chunk');
-      await apos.http.post(url, {
+      await http.post(url, {
         qs: {
           aposBigUpload: {
             type: 'chunk',
@@ -75,14 +83,12 @@ export default async (url, options) => {
         },
         body: formData
       });
-      console.log('after chunk');
       sentBytes += thisChunkSize;
       progress(sentBytes / totalBytes);
       chunk++;
     }
     n++;
   }
-  console.log('before end');
   const result = await http.post(url, {
     qs: {
       aposBigUpload: {
@@ -93,7 +99,5 @@ export default async (url, options) => {
     },
     body: options.body
   });
-  console.log('after end');
-  console.log(result);
   return result;
-}
+};
